requirements.txt
django==1.8.3
pillow==2.9.0
django-pure-pagination